Import SATF Library Sequence (#1592)
* Bugfix allow for paths in Headers

* Library Sequence Modal

This modal allows you to select a parcel and import a library sequence file

* Hook up the library modal to workspaces

* Fixed incorrect linting errors when multiple ranges are defined.

The linter incorrectly reported errors for values within a valid range when multiple ranges were defined. For example, if the ranges were 5-10 and 11-20, the linter would erroneously flag the value 8 as invalid because it falls outside the 11-20 range, even though it is valid within the 5-10 range.

* Improved autocomplete functionality by allowing completion even with invalid variable types.

Allow autocomplete to succeed and show the ERROR parameter value. The user will see this error in the editor and manually have to fix it.

* Fix @MODEL generation.

There shouldn't be any parenthesis and commas in the SeqN generation

* fix buttons wrapping

---------

Co-authored-by: bduran <bryan.duran@jpl.nasa.gov>